+++ tests/BaseTest.php
@@ -104,12 +104,24 @@ abstract class BaseTest extends TestCase {
         $config = [
             API::URL => getenv('URL'),
             API::KEY => getenv('KEY'),
-            API::LANGUAGE => getenv('IDOIT_LANGUAGE')
+            API::BYPASS_SECURE_CONNECTION => getenv('BYPASS_SECURE_CONNECTION')
         ];
+
+        if (getenv('IDOIT_LANGUAGE') !== false) {
+            $config[API::LANGUAGE] = getenv('IDOIT_LANGUAGE');
+        }
 
         if (getenv('USERNAME') !== false && getenv('PASSWORD') !== false) {
             $config[API::USERNAME] = getenv('USERNAME');
             $config[API::PASSWORD] = getenv('PASSWORD');
+        }
+
+        if (getenv('BYPASS_SECURE_CONNECTION') !== false) {
+            $config[API::BYPASS_SECURE_CONNECTION] = filter_var(
+                getenv('BYPASS_SECURE_CONNECTION'),
+                FILTER_VALIDATE_BOOLEAN,
+                FILTER_NULL_ON_FAILURE
+            );
         }
 
         $this->api = new API($config);